requirements.txt
requests
pandas
yfinance
numpy
matplotlib
scikit-learn
nltk
newsapi-python
transformers
reportlab
torch
tensorflow